+++ src/router/router.js
@@ -8,6 +8,7 @@ var getUser = require('../control/getUser');
 var getPost = require('../control/getPost');
 var getComment = require('../control/getComment');
 var postComment = require('../control/postComment');
+var commentLike = require('../control/commentLike');
 
 var router = express.Router();
 
@@ -15,30 +16,7 @@ router.get('/:address', getUser)
 router.get('/:address/:id', getPost)
 router.get('/:address/:id/comment', getComment)
 router.post('/:address/:id/comment', postComment)
-
-
-router.get('/:address/:postID/comment/like/:comID', (req, res) => {
-	fs.readFile("./data/comment.json", "utf8", (err, data) => {
-		if (err) {
-			
-			console.log("File read failed:", err);
-			res.send("error");
-			return;
-		}
-		var ret = JSON.parse(data);
-		var com = ret.data.find(comm => comm["comment-id"] === req.params.comID);
-		com["like-num"]++ ;
-		console.log(ret);
-		var json = JSON.stringify(ret); //convert it back to json
-		fs.writeFile('./data/comment.json', json, 'utf8', (err) => {
-			if (err){
-				console.log("write error");
-			}
-		});
-
-	})
-	res.send("comment like button");
-})
+router.get('/:address/:postID/comment/like/:comID', commentLike)
 
 
 module.exports = router;
